Add test configuration and setup
