+++ tests/integration/test_predefined_roles.py
@@ -1,265 +1,129 @@
 #!/usr/bin/env python3
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
-
-import asyncio
 import logging
 
-import psycopg2
-import psycopg2.sql
-import pytest
+import psycopg2 as psycopg2
+import pytest as pytest
+import yaml
 from pytest_operator.plugin import OpsTest
 from tenacity import Retrying, stop_after_delay, wait_fixed
 
 from .helpers import (
-    CHARM_BASE,
     DATA_INTEGRATOR_APP_NAME,
     DATABASE_APP_NAME,
     db_connect,
     get_password,
     get_primary,
     get_unit_address,
-)
-from .new_relations.helpers import (
-    build_connection_string,
-    get_application_relation_data,
-    get_juju_secret,
+    relations,
 )
 
 logger = logging.getLogger(__name__)
 
-TIMEOUT = 15 * 60
+REQUESTED_DATABASE_NAME = "requested-database"
+OTHER_DATABASE_NAME = "other-database"
+RELATION_ENDPOINT = "postgresql"
+CONFIGURATION_FILE_PATH = "tests/integration/predefined_roles.yaml"
 
-# NOTE: We are unable to test set_user_u('operator') as dba user because psycopg2
-# runs every query in a transaction and running set_user() is not supported in transactions.
+
+class PredefinedRole(yaml.YAMLObject):
+    yaml_tag = "!PredefinedRole"
+
+    def __init__(self, name, database_owner, in_roles, permissions):
+        self.name = name
+        self.database_owner = database_owner
+        self.in_roles = in_roles
+        self.permissions = permissions
+
+
+@pytest.fixture(scope="module")
+def predefined_roles() -> str:
+    with open(CONFIGURATION_FILE_PATH) as file:
+        data = yaml.load(file, Loader=yaml.Loader)
+        return data["extra-user-roles"]
+
+
+@pytest.fixture(scope="module")
+def predefined_roles_combinations() -> str:
+    with open(CONFIGURATION_FILE_PATH) as file:
+        data = yaml.load(file, Loader=yaml.Loader)
+        return data["allowed-combinations"]
 
 
 @pytest.mark.abort_on_fail
-@pytest.mark.skip_if_deployed
-async def test_deploy(ops_test: OpsTest, charm: str):
-    """Deploy the postgresql charm."""
-    async with ops_test.fast_forward("10s"):
-        await asyncio.gather(
-            ops_test.model.deploy(
-                charm,
-                application_name=DATABASE_APP_NAME,
-                num_units=2,
-                base=CHARM_BASE,
-                config={"profile": "testing"},
-            ),
-            ops_test.model.deploy(
+async def test_deploy(ops_test: OpsTest, charm, predefined_roles_combinations) -> None:
+    """Deploy and relate the charms."""
+    drop_databases = False
+    reset_relation = False
+
+    # Deploy the database charm if not already deployed.
+    if DATABASE_APP_NAME not in ops_test.model.applications:
+        logger.info("Deploying database charm")
+        await ops_test.model.deploy(charm, config={"profile": "testing"}, num_units=2)
+    else:
+        drop_databases = True
+    applications = [DATABASE_APP_NAME]
+
+    for index, combination in enumerate(predefined_roles_combinations):
+        # Drop the database requested by each data integrator when restarting the test.
+        database_name = f"{REQUESTED_DATABASE_NAME}-{index}"
+        if drop_databases:
+            logger.info(f"Dropping {database_name} database from already deployed database charm")
+            primary = await get_primary(ops_test, f"{DATABASE_APP_NAME}/0")
+            connection = None
+            try:
+                host = get_unit_address(ops_test, primary)
+                password = await get_password(ops_test, database_app_name=DATABASE_APP_NAME)
+                connection = db_connect(host, password)
+                connection.autocommit = True
+                with connection.cursor() as cursor:
+                    cursor.execute(f'DROP DATABASE IF EXISTS "{database_name}";')
+            finally:
+                if connection is not None:
+                    connection.close()
+            reset_relation = True
+
+        # Deploy the data integrator charm for each combination of predefined roles.
+        data_integrator_app_name = f"{DATA_INTEGRATOR_APP_NAME}{index}"
+        extra_user_roles = ",".join(combination)
+        if data_integrator_app_name not in ops_test.model.applications:
+            logger.info(
+                f"Deploying data integrator charm {'with extra user roles: ' + extra_user_roles.replace(',', ', ') if extra_user_roles else 'without extra user roles'}"
+            )
+            await ops_test.model.deploy(
                 DATA_INTEGRATOR_APP_NAME,
-                base=CHARM_BASE,
-            ),
-            ops_test.model.deploy(
-                DATA_INTEGRATOR_APP_NAME,
-                application_name=f"{DATA_INTEGRATOR_APP_NAME}2",
-                base=CHARM_BASE,
-            ),
-        )
-
-        await ops_test.model.wait_for_idle(
-            apps=[DATABASE_APP_NAME], status="active", timeout=TIMEOUT
-        )
-        assert ops_test.model.applications[DATABASE_APP_NAME].units[0].workload_status == "active"
-
-        await ops_test.model.wait_for_idle(
-            apps=[DATA_INTEGRATOR_APP_NAME, f"{DATA_INTEGRATOR_APP_NAME}2"],
-            status="blocked",
-            timeout=(5 * 60),
-        )
-
-
-@pytest.mark.abort_on_fail
-async def test_charmed_read_role(ops_test: OpsTest):
-    """Test the charmed_read predefined role."""
-    await ops_test.model.applications[DATA_INTEGRATOR_APP_NAME].set_config({
-        "database-name": "charmed_read_database",
-        "extra-user-roles": "charmed_read",
-    })
-    await ops_test.model.add_relation(DATA_INTEGRATOR_APP_NAME, DATABASE_APP_NAME)
-    await ops_test.model.wait_for_idle(
-        apps=[DATA_INTEGRATOR_APP_NAME, DATABASE_APP_NAME], status="active"
-    )
-
-    primary = await get_primary(ops_test, f"{DATABASE_APP_NAME}/0")
-    primary_address = get_unit_address(ops_test, primary)
-    operator_password = await get_password(ops_test, "operator")
-
-    with db_connect(
-        primary_address, operator_password, username="operator", database="charmed_read_database"
-    ) as connection:
-        connection.autocommit = True
-
-        with connection.cursor() as cursor:
-            cursor.execute("CREATE TABLE test_table (id SERIAL PRIMARY KEY, data TEXT);")
-            cursor.execute("INSERT INTO test_table (data) VALUES ('test_data'), ('test_data_2');")
-
-    connection_string = await build_connection_string(
-        ops_test,
-        DATA_INTEGRATOR_APP_NAME,
-        "postgresql",
-        database="charmed_read_database",
-    )
-
-    with psycopg2.connect(connection_string) as connection:
-        connection.autocommit = True
-
-        with connection.cursor() as cursor:
-            logger.info("Checking that the charmed_read role can read from the database")
-            cursor.execute(
-                "SELECT table_name FROM information_schema.tables WHERE table_name NOT LIKE 'pg_%' AND table_name NOT LIKE 'sql_%' AND table_type <> 'VIEW';"
+                application_name=data_integrator_app_name,
+                config={"database-name": database_name, "extra-user-roles": extra_user_roles},
             )
-            tables = [row[0] for row in cursor.fetchall()]
-            assert tables == ["test_table"], "Unexpected tables in the database"
+        else:
+            logger.info("Resetting extra user roles in already deployed data integrator charm")
+            await ops_test.model.applications[data_integrator_app_name].set_config({
+                "database-name": database_name,
+                "extra-user-roles": extra_user_roles,
+            })
+            reset_relation = True
 
-            cursor.execute("SELECT data FROM test_table;")
-            data = sorted([row[0] for row in cursor.fetchall()])
-            assert data == sorted(["test_data", "test_data_2"]), (
-                "Unexpected data in charmed_read_database with charmed_read role"
+        # Relate the data integrator charm to the database charm.
+        existing_relations = relations(ops_test, DATABASE_APP_NAME, data_integrator_app_name)
+        if reset_relation and existing_relations:
+            logger.info("Removing existing relation between charms")
+            await ops_test.model.applications[data_integrator_app_name].remove_relation(
+                f"{data_integrator_app_name}:{RELATION_ENDPOINT}", DATABASE_APP_NAME
             )
-            logger.info("Checking that the charmed_read role cannot create a new table")
-            with pytest.raises(psycopg2.errors.InsufficientPrivilege):
-                cursor.execute("CREATE TABLE test_table_2 (id INTEGER);")
-    connection.close()
-
-    with psycopg2.connect(connection_string) as connection, connection.cursor() as cursor:
-        logger.info("Checking that the charmed_read role cannot write to an existing table")
-        with pytest.raises(psycopg2.errors.InsufficientPrivilege):
-            cursor.execute(
-                "INSERT INTO test_table (data) VALUES ('test_data_3'), ('test_data_4');"
-            )
-    connection.close()
-
-    with psycopg2.connect(connection_string) as connection, connection.cursor() as cursor:
-        logger.info(
-            "Checking that the charmed_read role can't escalate to the database owner user"
-        )
-        with pytest.raises(psycopg2.errors.InsufficientPrivilege):
-            cursor.execute("SET ROLE charmed_read_database_owner;")
-    connection.close()
-
-    await ops_test.model.applications[DATABASE_APP_NAME].remove_relation(
-        f"{DATABASE_APP_NAME}:database", f"{DATA_INTEGRATOR_APP_NAME}:postgresql"
-    )
-    await ops_test.model.wait_for_idle(apps=[DATA_INTEGRATOR_APP_NAME], status="blocked")
-
-
-@pytest.mark.abort_on_fail
-async def test_charmed_dml_role(ops_test: OpsTest):
-    """Test the charmed_dml role."""
-    await ops_test.model.applications[DATA_INTEGRATOR_APP_NAME].set_config({
-        "database-name": "charmed_dml_database",
-        "extra-user-roles": "charmed_dml",
-    })
-    await ops_test.model.add_relation(DATA_INTEGRATOR_APP_NAME, DATABASE_APP_NAME)
-    await ops_test.model.wait_for_idle(
-        apps=[DATA_INTEGRATOR_APP_NAME, DATABASE_APP_NAME], status="active"
-    )
-
-    await ops_test.model.applications[f"{DATA_INTEGRATOR_APP_NAME}2"].set_config({
-        "database-name": "throwaway",
-        "extra-user-roles": "charmed_dml",
-    })
-    await ops_test.model.add_relation(f"{DATA_INTEGRATOR_APP_NAME}2", DATABASE_APP_NAME)
-    await ops_test.model.wait_for_idle(apps=[f"{DATA_INTEGRATOR_APP_NAME}2"], status="active")
-
-    primary = await get_primary(ops_test, f"{DATABASE_APP_NAME}/0")
-    primary_address = get_unit_address(ops_test, primary)
-    operator_password = await get_password(ops_test, "operator")
-
-    with db_connect(
-        primary_address, operator_password, username="operator", database="charmed_dml_database"
-    ) as connection:
-        connection.autocommit = True
-        with connection.cursor() as cursor:
-            cursor.execute("CREATE TABLE test_table (id SERIAL PRIMARY KEY, data TEXT);")
-    connection.close()
-
-    connection_string = await build_connection_string(
-        ops_test,
-        DATA_INTEGRATOR_APP_NAME,
-        "postgresql",
-        database="charmed_dml_database",
-    )
-
-    with psycopg2.connect(connection_string) as connection:
-        connection.autocommit = True
-
-        with connection.cursor() as cursor:
-            logger.info("Checking that the charmed_dml role can write data")
-            cursor.execute("INSERT INTO test_table (data) VALUES ('test_data'), ('test_data_2');")
-            logger.info("Checking that the charmed_dml role can read data")
-            cursor.execute("SELECT data FROM test_table;")
-            data = sorted([row[0] for row in cursor.fetchall()])
-            assert data == sorted(["test_data", "test_data_2"]), (
-                "Unexpected data in charmed_dml_database with charmed_dml role"
-            )
-    connection.close()
-
-    with (
-        psycopg2.connect(connection_string) as connection,
-        connection.cursor() as cursor,
-    ):
-        logger.info("Checking that the charmed_dml role can't escalate to the database owner user")
-        with pytest.raises(psycopg2.errors.InsufficientPrivilege):
-            cursor.execute("SET ROLE charmed_read_database_owner;")
-    connection.close()
-
-    secret_uri = await get_application_relation_data(
-        ops_test,
-        f"{DATA_INTEGRATOR_APP_NAME}2",
-        "postgresql",
-        "secret-user",
-    )
-    secret_data = await get_juju_secret(ops_test, secret_uri)
-    data_integrator_2_user = secret_data["username"]
-    data_integrator_2_password = secret_data["password"]
-
-    with db_connect(primary_address, operator_password, username="operator") as connection:
-        connection.autocommit = True
-
-        with connection.cursor() as cursor:
-            cursor.execute(
-                psycopg2.sql.SQL("GRANT connect ON DATABASE charmed_dml_database TO {};").format(
-                    psycopg2.sql.Identifier(data_integrator_2_user)
+            async with ops_test.fast_forward():
+                await ops_test.model.wait_for_idle(
+                    apps=[data_integrator_app_name], status="blocked"
                 )
-            )
-    connection.close()
+            logger.info("Adding relation between charms")
+            for attempt in Retrying(stop=stop_after_delay(60), wait=wait_fixed(3)):
+                with attempt:
+                    await ops_test.model.relate(data_integrator_app_name, DATABASE_APP_NAME)
+        if not existing_relations:
+            logger.info("Adding relation between charms")
+            await ops_test.model.relate(data_integrator_app_name, DATABASE_APP_NAME)
 
-    connection.close()
-    for attempt in Retrying(stop=stop_after_delay(60), wait=wait_fixed(3)):
-        with attempt:
-            connection = db_connect(
-                primary_address,
-                data_integrator_2_password,
-                username=data_integrator_2_user,
-                database="charmed_dml_database",
-            )
-    with connection.cursor() as cursor:
-        connection.autocommit = True
-        cursor.execute("INSERT INTO test_table (data) VALUES ('test_data_3');")
-    connection.close()
+        applications.append(data_integrator_app_name)
 
-    with db_connect(
-        primary_address, operator_password, username="operator", database="charmed_dml_database"
-    ) as connection:
-        connection.autocommit = True
-
-        with connection.cursor() as cursor:
-            cursor.execute("SELECT data FROM test_table;")
-            data = sorted([row[0] for row in cursor.fetchall()])
-            assert data == sorted(["test_data", "test_data_2", "test_data_3"]), (
-                "Unexpected data in charmed_read_database with charmed_read role"
-            )
-    connection.close()
-
-    await ops_test.model.applications[DATABASE_APP_NAME].remove_relation(
-        f"{DATABASE_APP_NAME}:database", f"{DATA_INTEGRATOR_APP_NAME}:postgresql"
-    )
-    await ops_test.model.applications[DATABASE_APP_NAME].remove_relation(
-        f"{DATABASE_APP_NAME}:database", f"{DATA_INTEGRATOR_APP_NAME}2:postgresql"
-    )
-    await ops_test.model.wait_for_idle(
-        apps=[DATA_INTEGRATOR_APP_NAME, f"{DATA_INTEGRATOR_APP_NAME}2"], status="blocked"
-    )
+    async with ops_test.fast_forward():
+        await ops_test.model.wait_for_idle(apps=applications, status="active")
