+++ tests/spread/test_predefined_instance_roles.py/task.yaml
@@ -1,0 +1,7 @@
+summary: test_predefined_instance_roles.py
+environment:
+  TEST_MODULE: test_predefined_instance_roles.py
+execute: |
+  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
+artifacts:
+  - allure-results
